Updates homepage UI for improved responsiveness
Improves homepage responsiveness, especially for larger screens.

Adjusts the QR code section in the header to be thinner.

Fixes an issue where the FeatureShortcuts component was not being dynamically imported correctly.

Ensures the status parameter is included in the getMediasHomepage API call.

Renders GridBanner conditionally to avoid errors when rightSlides is undefined.